Address large literal chunking review
